✨ feat: proto ghoster
feat(protoutils): add ghoster utility for masking proto fields  Introduced a new utility package `ghoster` to handle the masking of sensitive fields in protobuf messages. This includes the implementation of the `Secrets` function, which replaces or clears fields marked with the `(ghoster.secret)` option based on the `nuller` flag. The addition also includes corresponding tests, proto definitions, and generated code. This utility is useful for ensuring sensitive data is not exposed in logs or other outputs.